\name{hdfsReadLines}
\alias{hdfsReadLines}
\title{Reads all lines from a text file located on the HDFS}
\usage{
  hdfsReadLines(inp, l)
}
\arguments{
  \item{inp}{The location of the text file, interolated
  based on hdfs.getwd}
}
\description{
  Reads all lines from a text file located on the HDFS
}
\keyword{HDFS}
\keyword{TextFile}

