\name{Rhipe-package}
\alias{Rhipe-package}
\alias{Rhipe}
\alias{rhipe}
\docType{package}




\title{RHIPE: R and Hadoop Integrated Processing Environment}
\description{
  
  Integrates R with Hadoop
}

\details{
See the website for detailed documentation:	
  (\url{http://ml.stat.purdue.edu/rhipe})
  
See \code{\link{rhmr}} for MapReduce using Rhipe.

}
\author{
Saptarshi Guha  <sguha@purdue.edu>

Parts of documentation contributed by:

Jeremiah Rounds <jrounds@purdue.edu>

}


\keyword{ package }


