\name{RHIPE}
\alias{rhmr}
\alias{rhlapply}
\alias{rhread}
\alias{rhreadBin}
\alias{rhsz}
\alias{rhuz}
\alias{rhmerge}
\alias{rhput}
\alias{rhget}

\title{RHIPE: R and Hadoop Integrated Processing Environment}
\description{
  
  Integrates R with Hadoop
}

\details{
	See the website	
  (\url{http://ml.stat.purdue.edu/rhipe})

}
