\name{rhJobInfo}
\alias{rhJobInfo}
\title{Very Detailed Job Information}
\usage{
  rhJobInfo(job)
}
\arguments{
  \item{job}{The parameter \code{job} can either be a
  string with the format \emph{job_datetime_id} (e.g.
  \emph{job_201007281701_0274}) or the value returned from
  \code{rhex} with the \code{async} option set to TRUE.}
}
\description{
  Returns detailed job information across all TaskIDs
}

