\name{hdfsReadLines}
\alias{hdfsReadLines}
\title{Reads all or some lines from a text file located on the HDFS}
\usage{
  hdfsReadLines(inp, l = -1L)
}
\arguments{
  \item{inp}{The location of the text file, interolated
  based on hdfs.getwd}

  \item{l}{the number of lines to read}
}
\description{
  Reads all or some lines from a text file located on the
  HDFS
}
\keyword{HDFS}
\keyword{TextFile}

